+++ database/migrations/2016_10_19_084045_add_resource_qty_manual.php
@@ -13,7 +13,7 @@ class AddResourceQtyManual extends Migration
     public function up()
     {
         Schema::table('breakdown_resources', function (Blueprint $table) {
-            $table->float('resource_qty')->nullable();
+            $table->double('resource_qty',12,2)->nullable();
             $table->boolean('resource_qty_manual')->nullable();
         });
     }
